% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_citation_matching.R
\name{abbreviate_term}
\alias{abbreviate_term}
\title{Abbreviate a single term using LTWA}
\usage{
abbreviate_term(word, ltwa_lookup, common_abbr, check = TRUE)
}
\arguments{
\item{word}{Single word to abbreviate}

\item{ltwa_lookup}{Pre-processed LTWA lookup tables}

\item{common_abbr}{Named vector of common abbreviations}

\item{check}{Logical, whether to check for abbreviation}
}
\value{
Abbreviated form of word
}
\description{
Abbreviate a single term using LTWA
}
\keyword{internal}
