% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblio.R
\docType{data}
\name{biblio}
\alias{biblio}
\title{Dataset of "Bibliometrics" manuscripts.}
\format{A large character with 9014 rows.\cr 
Data has been imported by an ISI Export file in bibtex format using the function \code{\link{readLines}}.}
\source{
\url{http://www.webofknowledge.com}
}
\usage{
biblio
}
\description{
The set of manuscripts which the title containing the word "bibliometrics" and published in a journal indexed by ISI WoK database.\cr
Period: 2006 - 2015\cr
Database: \href{http://www.webofknowledge.com}{ISI Web of Knowledge}\cr
}
\keyword{datasets}

