% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtag.R
\name{bibtag}
\alias{bibtag}
\title{Tag list and bibtex fields.}
\format{A data frame with 39 rows and 4 variables:
 \describe{
    \item{TAG}{country names}
    \item{SCOPUS}{continent names}
    \item{ISI}{country centroid longitude}
    \item{GENERIC}{country centroid latitude}
    }}
\description{
Data frame containing a list of tags and corresponding WoS and SCOPUS bibtex field.
}
