% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\docType{data}
\name{countries}
\alias{countries}
\title{Index of Countries.}
\format{A data frame with 198 rows and 1 variable:
 \describe{
    \item{countries}{country names}
    }}
\usage{
countries
}
\description{
Data frame containing a normalized index of countries.\cr
Data are used by \code{\link{biblioAnalysis}} function 
to extract Country Field of Cited References and Authors.
}
\keyword{datasets}

