% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customTheme.R
\name{customTheme}
\alias{customTheme}
\title{Custom Theme variables for Biblioshiny.}
\format{
A list with 3 elements:
 \describe{
    \item{name}{object name}
    \item{attribs}{attributes}
    \item{children}{CSS style}
    }
}
\description{
List containing a set of custom theme variables for Biblioshiny.
}
