% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garfield.R
\name{garfield}
\alias{garfield}
\title{Eugene Garfield's manuscripts.}
\format{
A data frame with 147 rows and 15 variables:
\describe{
\item{AU}{Authors}
\item{TI}{Document Title}
\item{SO}{Publication Name (or Source)}
\item{JI}{ISO Source Abbreviation}
\item{DT}{Document Type}
\item{DE}{Author Keywords}
\item{ID}{Keywords associated by WoS or SCOPUS database}
\item{AB}{Abstract}
\item{C1}{Author Address}
\item{RP}{Reprint Address}
\item{CR}{Cited References}
\item{TC}{Times Cited}
\item{PY}{Year}
\item{UT}{Unique Article Identifier}
\item{DB}{Database}
}
}
\source{
\url{https://www.scopus.com}
}
\description{
All manuscripts published by Eugene Garfield.\cr
Period: 1954 - 2014\cr
Database: \href{https://www.scopus.com}{SCOPUS source}\cr
}
