% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_citation_matching.R
\name{get_iso4_stop_words}
\alias{get_iso4_stop_words}
\title{Articles, prepositions, and conjunctions to be removed (ISO 4 standard)}
\usage{
get_iso4_stop_words()
}
\description{
Articles, prepositions, and conjunctions to be removed (ISO 4 standard)
}
\keyword{internal}
