% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeCycle.R
\name{lifeCycle}
\alias{lifeCycle}
\title{Life Cycle Analysis with Logistic Growth Model}
\usage{
lifeCycle(data, forecast_years = 5, plot = TRUE, verbose = FALSE)
}
\arguments{
\item{data}{Data frame with columns: year (PY) and number of publications (n)}

\item{forecast_years}{Number of years to forecast beyond saturation}

\item{plot}{Logical, if TRUE produces plots}

\item{verbose}{Logical, if TRUE prints detailed output}
}
\value{
List containing parameters, forecasts and metrics
}
\description{
Estimates logistic growth model for annual (non-cumulative) publications
following Meyer et al. (1999) methodology
}
