% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltwa.R
\name{ltwa}
\alias{ltwa}
\title{Index of ltwa.}
\format{
A data frame with 56463 rows and 3 variables:
 \describe{
    \item{WORD}{word from journal names}
    \item{ABBREVIATION}{ISO4 abbreviation}
    \item{LANGUAGES}{Language of the journal name}
    }
}
\description{
Data frame containing a normalized index of words used in journal names and their ISO4 abbreviations.
}
