% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{management}
\alias{management}
\title{The use of bibliometric approaches in business and management disciplines.}
\format{
A data frame with 449 rows and 66 variables:
\describe{
\item{PT}{Publication Type (J=Journal; B=Book; S=Series; P=Patent)}
\item{AU}{Author names}
\item{AF}{Author fullnames}
\item{TI}{Document Title}
\item{SO}{Publication Name (or Source)}
\item{LA}{Language}
\item{DT}{Document Type}
\item{DE}{Author Keywords}
\item{ID}{Keywords associated by ISI or SCOPUS database}
\item{AB}{Abstract}
\item{C1}{Author Address}
\item{RP}{Reprint Address}
\item{EM}{Email}
\item{FU}{Funding Agency and Grant Number}
\item{FX}{Funding Text}
\item{CR}{Cited References}
\item{NR}{Cited Reference Count}
\item{TC}{WoS Times Cited}
\item{Z9}{Total Times Cited Count (All Databases)}
\item{U1}{Usage Count (Last 180 Days)}
\item{U2}{Usage Count (Since 2013)}
\item{PU}{Publisher}
\item{PI}{Publisher City}
\item{PA}{Publisher Address}
\item{SN}{ISSN}
\item{EI}{eISSN}
\item{J9}{29-Character Source Abbreviation}
\item{JI}{ISO Source Abbreviation}
\item{PD}{Publication Date}
\item{PY}{Year Published}
\item{VL}{Volume}
\item{IS}{Issue}
\item{BP}{Beginning Page}
\item{EP}{Ending Page}
\item{DI}{DOI}
\item{PG}{Page Count}
\item{WC}{Web of Science Categories}
\item{SC}{Research Areas}
\item{GA}{Document Delivery Number}
\item{UT}{Unique Article Identifier}
\item{DA}{Date this collection was downloaded}
\item{ER}{End of Record}
\item{OI}{ORCID Identifier}
\item{HC}{ESI Highly Cited Paper}
\item{HP}{ESI Hot Paper}
\item{RI}{ResearcherID Number}
\item{OA}{Open Access Indicator}
\item{SI}{Special Issue}
\item{AR}{Article Number}
\item{CT}{Conference Title}
\item{CY}{Conference Date}
\item{CL}{Conference Location}
\item{SP}{Conference Sponsors}
\item{PM}{PubMed ID}
\item{PN}{Part Number}
\item{SU}{Supplement}
\item{HO}{Conference Host}
\item{BE}{Editors}
\item{SE}{Book Series Title}
\item{BN}{ISBN}
\item{DB}{Database}
\item{AU_UN}{Author's Affiliations (disambiguated)}
\item{AU1_UN}{Corresponding Author's Affiliation (disambiguated)}
\item{AU_UN_NR}{Not Recognized Affiliations}
\item{SR_FULL}{Short Full-Reference}
\item{SR}{Short Reference}
}
}
\source{
\url{https://www.webofknowledge.com}
}
\description{
A collection of Articles about the use of bibliometric approaches in business and management disciplines\cr
Period: 1985 - 2018\cr
Database: \href{https://www.webofknowledge.com}{Clarivate Analytics Web of Science}\cr
}
