% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingData.R
\name{missingData}
\alias{missingData}
\title{Completeness of bibliographic metadata}
\usage{
missingData(df)
}
\arguments{
\item{M}{is a bibliographic data frame obtained by \code{\link{convert2df}} function.}
}
\value{
The function \code{missingData} returns a list containing two objects:
\tabular{lll}{
\code{allTags}  \tab   \tab is a data frame including results for all original metadata tags from the collection\cr
\code{mandatoryTags}\tab    \tab is a data frame that included only the tags needed for analysis with bibliometrix and biblioshiny.}
}
\description{
It calculates the percentage of missing data in the metadata of a bibliographic data frame.
}
\details{
Each metadata is assigned a status c("Excellent," "Good," "Acceptable," "Critical," "Completely missing") 
depending on the percentage of missing data ("0%," "1% to 10%," "11% to 20%"m "21% to 99%," "100%").

The results of the function allow us to understand which analyses can be performed with bibliometrix 
and which cannot based on the completeness (or status) of different metadata.
}
\examples{
data(scientometrics, package = "bibliometrixData")
res <- missingData(scientometrics)
print(res$mandatoryTags)

}
