% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bibliodendrogram.R
\name{plot.bibliodendrogram}
\alias{plot.bibliodendrogram}
\title{Plotting dendrogram resulting from Conceptual Structure Analysis}
\usage{
\method{plot}{bibliodendrogram}(x, ...)
}
\arguments{
\item{x}{is the object for which plots are desired.}

\item{...}{is a generic param for plot functions.}
}
\value{
The function \code{plot} draws a dendrogram.
}
\description{
\code{plot} method for class '\code{bibliodendrogram}'
}
