% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAuthorWorks.r
\name{print_author_works_summary}
\alias{print_author_works_summary}
\title{Print Summary of Author Works Search}
\usage{
print_author_works_summary(works_df)
}
\arguments{
\item{works_df}{Data.frame. Result from find_author_works()}
}
\description{
Prints a summary of the search results from find_author_works()
}
