% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{readFiles}
\alias{readFiles}
\title{Load a sequence of ISI or SCOPUS Export files into a large character object}
\usage{
readFiles(...)
}
\arguments{
\item{...}{is a sequence of names of files downaloaded from ISI WOS.(in plain text or bibtex format) or SCOPUS Export file (exclusively in bibtex format).}
}
\value{
a character vector of length the number of lines read.
}
\description{
It loads a sequence of SCOPUS and Thomson Reuters' ISI Web of Knowledge export files and create a large character vector from it.
}
\examples{
# ISI or SCOPUS Export files can be read using \\code{\\link{readFiles}} function:

# largechar <- readFiles('filename1.txt','filename2.txt','filename3.txt')

# filename1.txt, filename2.txt and filename3.txt are ISI or SCOPUS Export file 
# in plain text or bibtex format.

D <- readFiles('http://www.bibliometrix.org/datasets/bibliometrics_articles.txt')

}
\seealso{
\code{\link{convert2df}} for converting SCOPUS of ISI Export file into a dataframe
}

