% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scientometrics_text.R
\name{scientometrics_text}
\alias{scientometrics_text}
\title{"Co-citation analysis" and "Coupling analysis" manuscripts.}
\format{
A large character with 12731 rows.\cr 
Data has been imported by a WoS Export file in plain text format using the function \code{\link{readLines}}.
}
\source{
\url{https://www.webofknowledge.com}
}
\description{
Manuscripts about the topics "co-citation analysis" and 
"coupling analysis" published on Scientometrics Journal.\cr
Period: 1985 - 2015\cr
Database: \href{https://www.webofknowledge.com}{Clarivate Analytics Web of Science}\cr
}
