% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopusCollection.R
\name{scopusCollection}
\alias{scopusCollection}
\title{"Bibliometrics" manuscripts from SCOPUS.}
\format{
A data frame with 487 rows and 15 variables:
\describe{
\item{AU}{Authors}
\item{TI}{Document Title}
\item{SO}{Publication Name (or Source)}
\item{JI}{ISO Source Abbreviation}
\item{DT}{Document Type}
\item{DE}{Author Keywords}
\item{ID}{Keywords associated by WoS or SCOPUS database}
\item{AB}{Abstract}
\item{C1}{Author Address}
\item{RP}{Reprint Address}
\item{CR}{Cited References}
\item{TC}{Times Cited}
\item{PY}{Year}
\item{UT}{Unique Article Identifier}
\item{DB}{Database}
}
}
\source{
\url{https://www.scopus.com}
}
\description{
Manuscripts including the term "bibliometrics" in the title.\cr
Period: 1975 - 2017\cr
Database: \href{https://www.scopus.com}{SCOPUS}\cr
Format: bibtex\cr
}
