% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{stopwords}
\alias{stopwords}
\title{List of English stopwords.}
\format{
A character vector with 665 rows.\cr
}
\description{
A character vector containing a complete list of English stopwords\cr
Data are used by  \code{\link{biblioAnalysis}} function 
to extract Country Field of Cited References and Authors.
}
