% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblio.R
\name{biblio}
\alias{biblio}
\title{Dataset of "Bibliometrics" scientific documents.}
\format{
A large character with 9014 rows.\cr 
Data has been imported by an Clarivate Analytics WoS Export file in bibtex format using the function \code{\link{readLines}}.
}
\source{
\url{https://www.webofknowledge.com}
}
\description{
The set of manuscripts which the title containing the word "bibliometrics" and published in a journal indexed by Clarivate Analytics WoS database.\cr
Period: 2006 - 2015\cr
Database: \href{https://www.webofknowledge.com}{ISI Web of Science}\cr
}
