% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioshiny.R
\name{biblioshiny}
\alias{biblioshiny}
\title{Shiny UI for bibliometrix package}
\usage{
biblioshiny()
}
\description{
\code{biblioshiny} performs science mapping analysis using the main functions of the bibliometrix package.
}
\examples{

#biblioshiny()


}
