% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocMatrix.R
\name{cocMatrix}
\alias{cocMatrix}
\title{Co-occurrence matrix}
\usage{
cocMatrix(M, Field = "AU", type = "sparse", sep = ";")
}
\arguments{
\item{M}{is a data frame obtained by the converting function
\code{\link{convert2df}}. It is a data matrix with cases corresponding to
articles and variables to Field Tag in the original ISI or SCOPUS file.}

\item{Field}{is a character object. It indicates one of the field tags of the
  standard ISI WoS Field Tag codify. Field can be equal to one of this tags:
  \tabular{lll}{ \code{AU}\tab   \tab Authors\cr \code{SO}\tab   \tab
  Publication Name (or Source)\cr \code{JI}\tab   \tab ISO Source
  Abbreviation\cr \code{DE}\tab   \tab Author Keywords\cr \code{ID}\tab
  \tab Keywords associated by ISI or SCOPUS database \cr \code{CR}\tab   \tab
  Cited References}

  for a complete list of filed tags see:
  \href{https://images.webofknowledge.com/WOK50B6/help/WOS/h_fieldtags.html}{ISI
  WoS Field Tags}}

\item{type}{indicates the output format of co-occurrences: \tabular{lll}{
\code{type = "matrix"} \tab   \tab produces an object of class
\code{matrix}\cr \code{type = "sparse"} \tab   \tab produces an object of
class \code{dgMatrix} of the package \code{\link{Matrix}}. "sparse"
argument generates a compact representation of the matrix.}}

\item{sep}{is the field separator character. This character separates strings in each 
column of the data frame. The default is \code{sep = ";"}.}
}
\value{
a co-occurrence matrix with cases corresponding to manuscripts and variables to the
  objects extracted from the Tag \code{Field}.
}
\description{
\code{cocMatrix} computes co-occurences between elements of a Tag Field from a bibliographic data frame. Manuscript is the unit of analysis.
}
\details{
This co-occurrence matrix can be tranformed into a collection of compatible
networks. Through matrix multiplication you can obtain different networks. 
The fuction follows the approach proposed by Batagely and Cerinsek (2013).\cr\cr
}
\examples{
# EXAMPLE 1: Articles x Authors co-occurrence matrix

data(scientometrics)
WA <- cocMatrix(scientometrics, Field = "AU", type = "sparse", sep = ";")

# EXAMPLE 2: Articles x Cited References co-occurrence matrix

# data(scientometrics)

# WCR <- cocMatrix(scientometrics, Field = "CR", type = "sparse", sep = ";")

# EXAMPLE 3: Articles x Cited First Authors co-occurrence matrix

# data(scientometrics)
# scientometrics <- metaTagExtraction(scientometrics, Field = "CR_AU", sep = ";")
# WCR <- cocMatrix(scientometrics, Field = "CR_AU", type = "sparse", sep = ";")

}
\seealso{
\code{\link{convert2df}} to import and convert an ISI or SCOPUS
  Export file in a data frame.

\code{\link{biblioAnalysis}} to perform a bibliometric analysis.

\code{\link{biblioNetwork}} to compute a bibliographic network.
}

