% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{countries}
\alias{countries}
\title{Index of Countries.}
\format{
A data frame with 199 rows and 5 variables:
 \describe{
    \item{countries}{country names}
    \item{continent}{continent names}
    \item{iso2}{country ISO 3166-1 alpha-2 code}
    \item{Longitude}{country centroid longitude}
    \item{Latitude}{country centroid latitude}
    }
}
\description{
Data frame containing a normalized index of countries.\cr
Data are used by \code{\link{biblioAnalysis}} function
to extract Country Field of Cited References and Authors.
}
