% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isibib2df.R
\name{isibib2df}
\alias{isibib2df}
\title{Convert a Clarivate Analytics WoS Export file into a data frame}
\usage{
isibib2df(D)
}
\arguments{
\item{D}{is a character array containing data read from an WoS Export file (in bibtex format).}
}
\value{
a data frame with cases corresponding to articles and variables to Field Tag in the original SCOPUS file.
}
\description{
It is an internal function used by \code{\link{convert2df}} to convert a Clarivate Analytics WoS Export file 
and create a data frame from it, with cases corresponding to articles and variables to Field Tag in the original file.
}
\examples{

}
\seealso{
\code{\link{isi2df}} for converting ISI Export file (in plain text format)
}
\concept{converting functions}
