% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isiCollection.R
\name{isiCollection}
\alias{isiCollection}
\title{"Bibliometrics" manuscripts from Clarivate Analytics WoS.}
\format{
A data frame with 329 rows and 16 variables:
\describe{
\item{AU}{Authors}
\item{TI}{Document Title}
\item{SO}{Publication Name (or Source)}
\item{JI}{ISO Source Abbreviation}
\item{DT}{Document Type}
\item{DE}{Author Keywords}
\item{ID}{Keywords associated by WoS or SCOPUS database}
\item{AB}{Abstract}
\item{C1}{Author Address}
\item{RP}{Reprint Address}
\item{CR}{Cited References}
\item{TC}{Times Cited}
\item{PY}{Year}
\item{SC}{Subject Category}
\item{UT}{Unique Article Identifier}
\item{DB}{Database}
}
}
\source{
\url{https://www.webofknowledge.com}
}
\description{
Manuscripts including the term "bibliometrics" in the title.\cr
Period: 1985 - 2017\cr
Database: \href{https://www.webofknowledge.com}{Clarivate Analytics Web of Science}\cr
Format: bibtex\cr
}
